Keep Playwright Test Results
This patch makes Playwright take screenshots of failing tests and will
upload them alongside the test result report to the GitHub Actions
workflow artifacts. This makes it easier to get information about what
is failing if a test is failing.

This also limits the number of retries: If we have to retry three times
to pass a test something is definitely wrong. Either with the code or
with the test.